test(formatters): move formatter tests to dedicated directory
